+++ src/businessEntityType.ts
@@ -1,0 +1,3 @@
+import { ICategoryCode } from './categoryCode';
+
+export type IBusinessEntityType = Pick<ICategoryCode, 'codeValue' | 'typeOf'>;